+++ src/server/game/Battlegrounds/BattlegroundUtils.h
@@ -1,0 +1,8 @@
+#ifndef BATTLEGROUNDUTILS_H
+#define BATTLEGROUNDUTILS_H
+
+#include "Battleground.h"
+
+uint32 GetMinPlayersPerTeam(Battleground* bg, PvPDifficultyEntry const* bracketEntry);
+
+#endif  // BATTLEGROUNDUTILS_H
